+++ owlcms/src/main/java/com/github/appreciated/css/grid/exception/NegativeValueException.java
@@ -4,11 +4,15 @@ import java.util.Arrays;
 
 @SuppressWarnings("serial")
 public class NegativeValueException extends RuntimeException {
-    public NegativeValueException(int... value) {
-        super("One the passed values \"" + Arrays.stream(value).mapToObj(String::valueOf).reduce((left, right) -> left + ", " + right).orElse("") + "\" is negative. This Constructor requires the parameter to be positive");
-    }
+	public NegativeValueException(double... value) {
+		super("One the passed values \""
+		        + Arrays.stream(value).mapToObj(String::valueOf).reduce((left, right) -> left + ", " + right).orElse("")
+		        + "\" is negative. This Constructor requires the parameter to be positive");
+	}
 
-    public NegativeValueException(double... value) {
-        super("One the passed values \"" + Arrays.stream(value).mapToObj(String::valueOf).reduce((left, right) -> left + ", " + right).orElse("") + "\" is negative. This Constructor requires the parameter to be positive");
-    }
+	public NegativeValueException(int... value) {
+		super("One the passed values \""
+		        + Arrays.stream(value).mapToObj(String::valueOf).reduce((left, right) -> left + ", " + right).orElse("")
+		        + "\" is negative. This Constructor requires the parameter to be positive");
+	}
 }
